pyproject.toml
[project]
name = "house-agent-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.125.0",
    "google-api-python-client>=2.192.0",
    "google-auth>=2.47.0",
    "google-auth-oauthlib>=1.3.0",
    "gspread>=6.2.1",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "twilio>=9.10.3",
    "uvicorn>=0.38.0",
]
